🏗️ Switch PDF engine to XeLaTeX and change fonts
The math characters in the Agda code blocks were giving me trouble, as LaTeX could not process them.
So I tried switching to XeLaTeX, which has Unicode support out of the box.

I ended up switching to TeX Gyre for text and math, and DejaVu Sans Mono for mono, as it has all
necessary math characters so far.